serving/requirements.txt
Flask==2.0.1
numpy==1.24.3
Pillow==9.5.0
PyYAML==6.0
torch==1.9.0
torchvision==0.10.0
opencv-python==4.7.0.72
